// frontend/src/themes/colors.d.ts
declare const primary: {
  100: "#EEF6FC";
  200: "#97D5FC";
  300: "#47A8E5";
  400: "#2A6BAC";
  500: "#133A7C";
  600: "#081534";
};

declare const neutral: {
  0: "#FFFFFF";
  100: "#FAFAFA";
  200: "#F5F5F5";
  300: "#D1D4D5";
  400: "#848C8F";
  500: "#465357";
  600: "#232A2C";
};

declare const success: {
  medium: "#6ABB7B";
  light: "#D3FEDA";
};

declare const warning: {
  medium: "#DEC436";
  light: "#FFFFCF";
};

declare const error: {
  medium: "#B7424F";
  light: "#FFDDDF";
};

declare const background: {
  default: "#FAFAFA";
  paper: "#FFFFFF";
};

export { primary, success, warning, error, neutral, background };